requirements.txt
d2to1>=0.2.10,<0.3
eventlet
flask
iso8601
oslo.config
pbr>=0.6,<1.0
